requirements.txt
openai
typer
rich
pytest
coverage
pydantic
pyyaml
